% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_mean.R
\name{group_mean}
\alias{group_mean}
\title{Calculate group mean of numeric variable}
\usage{
group_mean(df, group, x, ..., na.rm = TRUE)
}
\arguments{
\item{df}{dataframe being used for analysis}

\item{group}{observational group which you would like to organize data by}

\item{x}{numeric variable which the mean will be taken from}

\item{...}{allows user to access additional arguments for the mean function}

\item{na.rm}{specifies that NA values will not be included in the mean calculation}
}
\value{
returns the mean of a numeric variable for each group of observations
}
\description{
Groups specified observations together and provides the mean of
a chosen numeric variable for each group
}
\details{
Groups specified observations together and provides the mean of a
chosen numeric variable for each group
}
\examples{
group_mean(penguins, species, body_mass_g)

group_mean(gapminder, continent, gdpPercap)

}
